web-animations/interfaces/Animatable/animate-pseudo-element.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animatable.animate pseudo element tests</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animatable-animate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<script src="../../resources/keyframe-utils.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

// Tests on CSSPseudoElement

test(function(t) {
  var pseudoTarget = createPseudo(t, 'before');
  var anim = pseudoTarget.animate(null);
  assert_class_string(anim, 'Animation', 'The returned object is an Animation');
}, 'CSSPseudoElement.animate() creates an Animation object');

test(function(t) {
  var pseudoTarget = createPseudo(t, 'before');
  var anim = pseudoTarget.animate(null);
  assert_equals(anim.effect.target, pseudoTarget,
                'The returned Animation targets to the correct object');
}, 'CSSPseudoElement.animate() creates an Animation object targeting ' +
   'to the correct CSSPseudoElement object');
</script>
</body>
